Exercise 6 - Brute Force Attack (Advance Requirements)
A program that simulates a password entry system. The correct password is defined and the program will keep asking the user to enter the password until they provide the correct one but there is a maximum number of 5 attempts, when this attempts is reached, the program will inform the user that the authorities have been alerted.